:bug: fix : 자료형으로 인한 계산식 오류 수정

+++ src/main/java/lotto/Domain/LottoResult.java
@@ -29,7 +29,7 @@ public class LottoResult {
 
     public String calculateSumOfReward(Money money) {
         double beforeMoney = money.getValue();
-        double sum = 0;
+        long sum = 0;
         for (LottoReward lottoReward : LottoReward.values()) {
             sum += lottoReward.getReward() * resultStatistic.getOrDefault(lottoReward, 0);
         }
